pyproject.toml
[tool.poetry]
name = "scrapybara-python-template"
version = "0.1.0"
description = ""
authors = ["nalin <74811063+lordspline@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
scrapybara = "2.1.2"
httpx = "0.27"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"